Task3 - CLU APPLICATION/CLU application with DJango/requirement.txt
azure-ai-language-conversations
azure-core
Django==3.2.8